requirements.txt
# local package
-e .

# external requirements
awscli
black
click
coverage
flake8
pre-commit
pytest
python-dotenv>=0.5.1
Sphinx
sphinx_rtd_theme
spotipy
tox-conda
